Rename the runtime files
